Fix Events Page
Transfer components to their own files
Fix spacing issues
Add hover effects
+ additional small changes
