+++ src/components/EventCard/EventCard.css
@@ -3,13 +3,12 @@
    flex-direction: column;
    width: 250px;
     background: #FFFFFF;
-    box-shadow: 1px 1px 18px rgba(0, 0, 0, 0.2);
+    box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.2);
     border-radius: 10px;
-    margin: 0 15px 20px;
 }
 
 .event_card:hover {
-    box-shadow: 1px 1px 25px rgba(0, 0, 0, 0.3);
+    box-shadow: 1px 1px 20px rgba(0, 0, 0, 0.3);
 }
 
 .content {
