+++ src/components/FilterEvents/FilterEvents.css
@@ -1,0 +1,25 @@
+.decorativeLine-thin {
+    width: 100%;
+    height: 1px;
+    background-color: #d1d1d1;
+    margin: 22px 0;
+}
+
+
+.filterEvents-container {
+    min-width: 300px;
+    width: 300px;
+    height: 100%;
+}
+
+
+.filterEvents-btnPrikazhi {
+    width: 100%;
+    font-weight: 500;
+    padding: 8px;
+}
+
+.filterEvents-btnPrikazhi:hover {
+    background-color: #785edc;
+    cursor: pointer;
+}
