+++ src/components/Heading/Heading.css
@@ -1,0 +1,23 @@
+.filterEvents-heading {
+    background-color: #9B85F7;
+    border-radius: 5px;
+    margin-bottom: 10px;
+    overflow: hidden;
+}
+
+.datePickLabel--labelContent {
+    display: flex;
+    color: white;
+    font-size: 24px;
+    font-weight: 500;
+    justify-content: center;
+    text-align: center;
+    margin: 10px;
+}
+
+.decorativeLine {
+    width: 100%;
+    height: 5px;
+    background-color: #FFC58F;
+}
+
