--- src/components/SortEvents/SortEvents.css
@@ -1,70 +1,0 @@
-.sortEvents-heading {
-    background-color: #9B85F7;
-    border-radius: 5px;
-    margin-bottom: 10px;
-}
-
-.decorativeLine {
-    width: 100%;
-    height: 5px;
-    background-color: #FFC58F;
-}
-
-.decorativeLine-thin {
-    width: 100%;
-    height: 1px;
-    background-color: #d1d1d1;
-    margin: 22px 0;
-}
-
-.datePickLabel--labelContent {
-    display: flex;
-    color: white;
-    font-size: 24px;
-    font-weight: 500;
-    padding-top: 12px;
-    justify-content: center;
-    text-align: center;
-    margin: 10px;
-}
-
-.sortEvents-container {
-    width: 280px;
-    height: 100%;
-}
-
-.layout-border {
-    padding: 10px 10px 20px;
-    border: 1px solid #d9d7d7;
-    border-radius: 4px;
-    background-color: #ededed;
-}
-#sortEvents-cityDropdown {
-    width: 100%;
-    border: 1px solid #BDBDBD;
-    background-color: white;
-    border-radius: 6px;
-    font-family: Roboto;
-    font-size: 17px;
-    font-weight: 300;
-    padding: 8px;
-    margin-top: 3px;
-}
-
-.sortEvents-btnPrikazhi {
-    width: 100%;
-    background-color: #9582E3;
-    color: white;
-    font-size: 20px;
-    font-family: Ubuntu;
-    font-weight: 500;
-    padding: 8px;
-    border-radius: 5px;
-    border: none;
-    box-shadow: 0px 2px 7px rgba(0, 0, 0, 0.25);
-}
-
-.sortEvents-btnPrikazhi:hover {
-    background-color: #785edc;
-    cursor: pointer;
-}
