+++ src/pages/Events/events.css
@@ -6,16 +6,20 @@
 }
 
 .event-list {
-    padding-top: 20px;
-    max-width: 870px;
-    max-height: 450px;
+    max-width: 830px;
+    height: fit-content;
     display: flex;
     flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
+    gap: 20px;
 }
 
 .event-page_content {
     display: flex;
     gap: 30px;
+}
+
+.layout-border {
+    border-radius: 4px;
+    background-color: #ededed;
+    padding: 20px;
 }
